move file to its own dir
