+++ code/config/messages.js
@@ -1,0 +1,5 @@
+const messages = [
+	'welcomeMsg' = 'Hello here & there'
+];
+
+export {messages};
