--- code/tsconfig.json
@@ -1,29 +1,0 @@
-{
-  "compileOnSave": false,
-  "compilerOptions": {
-		"experimentalDecorators": true,
-    "baseUrl": ".",
-    "incremental": false,
-    "noImplicitAny": true,
-    "forceConsistentCasingInFileNames": true,
-    "jsx": "react",
-    "moduleResolution": "Node",
-    "target": "ES2020",
-    "module": "CommonJS",
-    "skipLibCheck": false,
-    "allowSyntheticDefaultImports": true,
-    "esModuleInterop": true,
-    "isolatedModules": true,
-    "strictBindCallApply": true,
-    "lib": ["dom", "esnext"],
-    "noUnusedLocals": true
-  },
-  "exclude": ["dist", "**/dist", "node_modules", "**/node_modules", "**/setup-jest.ts"],
-  "ts-node": {
-    "transpileOnly": true,
-    "files": true,
-    "compilerOptions": {
-      "types": ["node"]
-    }
-  }
-}
