+++ main.js
@@ -1,0 +1,3 @@
+import { messages } from "./code/dev-config";
+
+window.onload(alert(messages.at[0]));
